feat(conversation): implement compact_conversation method in ConversationService
