+++ crates/forge_api/src/forge_api.rs
@@ -73,60 +73,14 @@ impl<F: Services + Infrastructure> API for ForgeAPI<F> {
         self.app.conversation_service().upsert(conversation).await
     }
 
-    /// Compacts the context of the main agent for the given conversation and
-    /// persists it. Returns metrics about the compaction (original vs.
-    /// compacted tokens and messages).
     async fn compact_conversation(
         &self,
         conversation_id: &ConversationId,
     ) -> anyhow::Result<CompactionResult> {
-        // Fetch the conversation
-        let mut conversation = self
-            .app
+        self.app
             .conversation_service()
-            .find(conversation_id)
-            .await?
-            .ok_or_else(|| anyhow::anyhow!("Conversation not found"))?;
-
-        // Identify the main agent and extract existing context
-        let main_agent_id = AgentId::new(Conversation::MAIN_AGENT_NAME);
-        let agent = conversation.get_agent(&main_agent_id)?;
-        let context = conversation
-            .state
-            .get(&main_agent_id)
-            .and_then(|s| s.context.clone())
-            .unwrap_or_default();
-
-        // Compute original metrics
-        let original_tokens = context.estimate_token_count() as usize;
-        let original_messages = context.messages.len();
-
-        // Perform compaction
-        let new_context = self
-            .app
-            .compaction_service()
-            .compact_context(agent, context.clone(), None)
-            .await?;
-
-        // Compute compacted metrics
-        let compacted_tokens = new_context.estimate_token_count() as usize;
-        let compacted_messages = new_context.messages.len();
-
-        // Persist the updated context
-        conversation
-            .state
-            .entry(main_agent_id.clone())
-            .or_default()
-            .context = Some(new_context.clone());
-        self.app.conversation_service().upsert(conversation).await?;
-
-        // Return metrics
-        Ok(CompactionResult::new(
-            original_tokens,
-            compacted_tokens,
-            original_messages,
-            compacted_messages,
-        ))
+            .compact_conversation(conversation_id)
+            .await
     }
 
     fn environment(&self) -> Environment {
