+++ crates/forge_services/src/conversation.rs
@@ -2,29 +2,34 @@ use std::collections::HashMap;
 use std::sync::Arc;
 
 use anyhow::{Context as AnyhowContext, Result};
-use forge_domain::{Conversation, ConversationId, ConversationService, Workflow};
+use forge_domain::{
+    AgentId, CompactionResult, CompactionService, Conversation, ConversationId,
+    ConversationService, Workflow,
+};
 use serde_json::Value;
 use tokio::sync::Mutex;
 
+/// Service for managing conversations, including creation, retrieval, and
+/// updates
 #[derive(Clone)]
-pub struct ForgeConversationService {
+pub struct ForgeConversationService<C> {
     workflows: Arc<Mutex<HashMap<ConversationId, Conversation>>>,
+    compaction_service: Arc<C>,
 }
 
-impl Default for ForgeConversationService {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl ForgeConversationService {
-    pub fn new() -> Self {
-        Self { workflows: Arc::new(Mutex::new(HashMap::new())) }
+impl<C: CompactionService> ForgeConversationService<C> {
+    /// Creates a new ForgeConversationService with the provided compaction
+    /// service
+    pub fn new(compaction_service: Arc<C>) -> Self {
+        Self {
+            workflows: Arc::new(Mutex::new(HashMap::new())),
+            compaction_service,
+        }
     }
 }
 
 #[async_trait::async_trait]
-impl ConversationService for ForgeConversationService {
+impl<C: CompactionService> ConversationService for ForgeConversationService<C> {
     async fn update<F, T>(&self, id: &ConversationId, f: F) -> Result<T>
     where
         F: FnOnce(&mut Conversation) -> T + Send,
@@ -69,5 +74,52 @@ impl ConversationService for ForgeConversationService {
 
     async fn delete_variable(&self, id: &ConversationId, key: &str) -> Result<bool> {
         self.update(id, |c| c.delete_variable(key)).await
+    }
+
+    async fn compact_conversation(&self, id: &ConversationId) -> Result<CompactionResult> {
+        // Fetch the conversation
+        let mut conversation = self
+            .find(id)
+            .await?
+            .ok_or_else(|| anyhow::anyhow!("Conversation not found"))?;
+
+        // Identify the main agent and extract existing context
+        let main_agent_id = AgentId::new(Conversation::MAIN_AGENT_NAME);
+        let agent = conversation.get_agent(&main_agent_id)?;
+        let context = conversation
+            .state
+            .get(&main_agent_id)
+            .and_then(|s| s.context.clone())
+            .unwrap_or_default();
+
+        // Compute original metrics
+        let original_tokens = context.estimate_token_count() as usize;
+        let original_messages = context.messages.len();
+
+        // Perform compaction
+        let new_context = self
+            .compaction_service
+            .compact_context(agent, context.clone(), None)
+            .await?;
+
+        // Compute compacted metrics
+        let compacted_tokens = new_context.estimate_token_count() as usize;
+        let compacted_messages = new_context.messages.len();
+
+        // Persist the updated context
+        conversation
+            .state
+            .entry(main_agent_id.clone())
+            .or_default()
+            .context = Some(new_context.clone());
+        self.upsert(conversation).await?;
+
+        // Return metrics
+        Ok(CompactionResult::new(
+            original_tokens,
+            compacted_tokens,
+            original_messages,
+            compacted_messages,
+        ))
     }
 }
